// src/types/Modulo.tsx
// src/types/Modulo.tsx
export enum Modulo {
    USUARIOS = "USUARIOS", // ruta /usuarios
    PRODUCTOS = "PRODUCTOS", // ruta /productos
    PRODUCCION = "PRODUCCION", // ruta /produccion
    STOCK = "STOCK", // ruta /stock
    PROVEEDORES = "PROVEEDORES", // ruta /Proveedores
    COMPRAS = "COMPRAS", // ruta /compras
    SEGUIMIENTO_PRODUCCION = "SEGUIMIENTO_PRODUCCION", // ruta /asistente_produccion
    TRANSACCIONES_ALMACE = "TRANSACCIONES_ALMACE", // ruta /recepcion_mprima
    ACTIVOS = "ACTIVOS", // ruta /Activos
    CONTABILIDAD = "CONTABILIDAD", //ruta /Contabilidad
    PERSONAL_PLANTA = "PERSONAL_PLANTA", // ruta /personal
    BINTELLIGENCE = "BINTELLIGENCE" // ruta Bintelligence
}